reload diary
треба все спочатку...
