+++ src/components/DiaryPages/Breakfast.jsx
@@ -1,0 +1,16 @@
+
+import React from "react";
+import TitlePeriod from "./TitlePeriod";
+import sprite from '../../assets/images/icons/icons.svg';
+import breakfastImg from '../../assets/images/meal-periods/breakfast.png';
+
+const Breakfast = () => {
+    return (
+        <>
+        <h1>Breakfast</h1>
+        <TitlePeriod />
+        </>
+    );
+    };
+
+export default Breakfast;
