--- src/components/DiaryPages/DiaryPage.jsx
@@ -1,4 +1,0 @@
-export { EditBtn } from './EditBtn.jsx';
-export { GoBackPage } from './GoBackPage.jsx';
-export { ModalDiaryBtn } from './ModalDiaryBtn.jsx';
-export { TitlePeriod } from './TitlePeriod.jsx';
