+++ src/components/DiaryPages/Dinner.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const Dinner = () => {
+    return (
+        <>
+        <h1>Dinner</h1>
+        </>
+    );
+    };
+
+export default Dinner;
