+++ src/components/DiaryPages/Lunch.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const Lunch = () => {
+    return (
+        <>
+        <h1>Lunch</h1>
+        </>
+    );
+    };
+
+export default Lunch;
