+++ src/components/DiaryPages/Snack.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+
+const Snack = () => {
+    return (
+        <>
+        <h1>Snack</h1>
+        </>
+    );
+    };
+
+export default Snack;
