--- src/components/DiaryPages/json/breakfast.json
@@ -1,34 +1,0 @@
-[
-  {
-    "name": "Apples",
-    "nutrition": {
-      "carbohydrates": 14,
-      "protein": 0.3,
-      "fat": 0.2
-    }
-  },
-  {
-    "name": "Strawberries",
-    "nutrition": {
-      "carbohydrates": 7.68,
-      "protein": 0.67,
-      "fat": 0.3
-    }
-  },
-  {
-    "name": "Blueberries",
-    "nutrition": {
-      "carbohydrates": 14.49,
-      "protein": 0.74,
-      "fat": 0.33
-    }
-  },
-  {
-    "name": "Spinach",
-    "nutrition": {
-      "carbohydrates": 3.6,
-      "protein": 2.9,
-      "fat": 0.4
-    }
-  }  
-]
