--- src/components/DiaryPages/json/dinner.json
@@ -1,42 +1,0 @@
-[
-  {
-    "name": "Kale",
-    "nutrition": {
-      "carbohydrates": 8.8,
-      "protein": 3.3,
-      "fat": 0.9
-    }
-  },
-  {
-    "name": "Quinoa",
-    "nutrition": {
-      "carbohydrates": 21.3,
-      "protein": 4.4,
-      "fat": 1.9
-    }
-  },
-  {
-    "name": "Greek Yogurt",
-    "nutrition": {
-      "carbohydrates": 3.6,
-      "protein": 10,
-      "fat": 0.4
-    }
-  },
-  {
-    "name": "Chicken Breast",
-    "nutrition": {
-      "carbohydrates": 0,
-      "protein": 31,
-      "fat": 3.6
-    }
-  },
-  {
-    "name": "Vegetable Soup",
-    "nutrition": {
-      "carbohydrates": 10,
-      "protein": 3,
-      "fat": 2
-    }
-  }
-]
