--- src/components/DiaryPages/json/lunch.json
@@ -1,26 +1,0 @@
-[  
-  {
-    "name": "Buckwheat",
-    "nutrition": {
-      "carbohydrates": 72,
-      "protein": 12,
-      "fat": 4
-    }
-  },
-  {
-    "name": "Eggs",
-    "nutrition": {
-      "carbohydrates": 1,
-      "protein": 13,
-      "fat": 11
-    }
-  },
-  {
-    "name": "Tofu",
-    "nutrition": {
-      "carbohydrates": 2,
-      "protein": 8,
-      "fat": 5
-    }
-  }  
-]
