--- src/components/DiaryPages/json/snack.json
@@ -1,18 +1,0 @@
-[
-  {
-    "name": "Berries",
-    "nutrition": {
-      "carbohydrates": 14.7,
-      "protein": 1.7,
-      "fat": 0.7
-    }
-  },
-  {
-    "name": "Dark leafy greens",
-    "nutrition": {
-      "carbohydrates": 4.7,
-      "protein": 4.3,
-      "fat": 0.4
-    }
-  }
-]
